1 create popups (#5)
* Rationalised data functionality

Now working but I want to rename "data" to records dataset=>recordset data=>records record=>as you were

Data is already in use for html data attributes in the layouts.

Other than that it's working pretty well.

Haven't actually done anything on the popups at the minute. The plan is to create a popup plug in and have open method on that. It should have a go() methods (go,next,previous).

* Popups first crack

see notes - two flavours for general and photos

Note that the javascript popup is for the photos and will need renaming.
 The javascript for standard popups is hard wired into the popup.html
 page and needs turning into a plug in.

* Some fixes for js stuff

* new docs