First working version of generic settings Edit
It works of a fashion. We have an object settingsTest which is a weird
combination of singleton pattern object and persistent store of data
This will get reworked to be purely singleton and accept all the data
by reference

First though we will turn this into a generic editing page that allows
us to edit any cs type.